credit_approval_system/requirements.txt
asgiref==3.8.1
Django==4.2.10

djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
redis==5.2.0
sqlparse==0.5.2
celery
